% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaspectrum-package.R
\docType{package}
\name{metaspectrum-package}
\alias{metaspectrum-package}
\title{metaspectrum: Brief overview of methods and general hints}
\description{
R package \bold{metaspectrum} provides ...
}
\details{
R package \bold{metaspectrum} ...

Type \code{help(package = "metaspectrum")} for a listing of R functions
available in \bold{metaspectrum}.

Type \code{citation("metaspectrum")} on how to cite \bold{metaspectrum} in
publications.

To report problems and bugs ...

The development version of \bold{metaspectrum} is available on GitHub
\url{https://github.com/littlecanargie/metaspectrum}.
}
\references{
Shih ...
Title.
\emph{Journal},
\bold{x}, p1--p2
}
\author{
Ming-Chieh Shih \email{mcshih@mx.nthu.edu.tw}
}
\keyword{package}
