% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nma_incon.R
\name{nma_incon}
\alias{nma_incon}
\title{Evidence-splitting model for network meta-analysis}
\usage{
nma_incon(
  s = "s",
  y = "y",
  v = "v",
  trt = "trt",
  comp,
  data = NULL,
  printout = TRUE,
  large_var = 1e+05,
  conf = 0.95
)
}
\arguments{
\item{s}{Column name for study (default "s")}

\item{y}{Column name for reported outcome (default "y")}

\item{v}{Column name for outcome variance (default "v")}

\item{trt}{Column name for treatment (default "trt"); s, y, v, trt
can be input as vectors directly if data is set to NULL}

\item{comp}{Length-2 vector listing the treatments in the contrast
of interest; effect is calculated as the second treatment - the
first treatment}

\item{data}{Input data frame (long form, i.e. one study-treatment
pair per row)}

\item{printout}{Boolean for printing out the results (default TRUE)}

\item{large_var}{Large variance for incidental parameter problem
(default 1e5)}

\item{conf}{confidence level of the confidence intervals (default
.95); confidence intervals are constructed via likelihood-based
approach}
}
\value{
List ...
}
\description{
Evidence-splitting model for network meta-analysis
}
\details{
Evidence-splitting model for network meta-analysis

!! Currently only support continuous outcomes !!

!! Check inconsistency identifiablity (van Valkenhoef, 2012) beforehand!!
}
\examples{
library("dplyr")
library("tidyr")

#### Dong2013 LABA-ICS compared to LABA ####

data(Dong2013, package = "netmeta")

Dong_new <- as_tibble(Dong2013) \%>\% 
  group_by(id) \%>\% 
  summarize(have_zero = (prod(death) == 0) | (prod(randomized - death) ==0), 
            all_zero = (sum(death) == 0) | (sum(randomized - death) ==0)) \%>\% 
  filter(!all_zero) \%>\%
  inner_join(Dong2013) \%>\%
  mutate(y = log((death + 0.5 * have_zero) / (randomized - death + 0.5 * have_zero)),
         v = 1/(death + 0.5 * have_zero) + 1/(randomized - death + 0.5 * have_zero))

Dong_res <- nma_incon(y = "y", v = "v", trt = "treatment", s = "id", 
                      comp = c("LABA", "LABA-ICS"), data = Dong_new)

#### Linde2015 Hypericum compared to Placebo ####

data(Linde2015, package = "netmeta")

Linde_new <- as_tibble(Linde2015) \%>\%
  select(id, starts_with("treatment"), starts_with("resp"), starts_with("n")) \%>\%
  pivot_longer(treatment1:n3, names_to = ".value", names_pattern = "(.*).") \%>\%
  filter(!is.na(resp))

Linde_new <- Linde_new \%>\% 
  group_by(id) \%>\% 
  summarize(have_zero = (prod(resp) == 0) | (prod(resp - n) ==0), 
            all_zero = (sum(resp) == 0) | (sum(resp - n) ==0)) \%>\% 
  filter(!all_zero) \%>\%
  inner_join(Linde_new) \%>\%
  mutate(y = log((resp + 0.5 * have_zero) / (n - resp + 0.5 * have_zero)),
         v = 1/(resp + 0.5 * have_zero) + 1/(n - resp + 0.5 * have_zero))

Linde_res <- nma_incon(y = "y", v = "v", trt = "treatment", s = "id", 
                      comp = c("Placebo", "Hypericum"), data = Linde_new)

### Pagliaro1992 (the data used in Shih & Tu, 2021) ###
### Beta-blocker compared to Sclerotherapy          ###

data(Pagliaro1992 , package = "meta")
Pagliaro_new <- as_tibble(Pagliaro1992) \%>\% 
  group_by(id) \%>\% 
  slice(1) \%>\% 
  ungroup() \%>\%
  select(id, bleed.plac, n.plac) \%>\% 
  rename(bleed.exp = bleed.plac, n.exp = n.plac) \%>\%
  mutate(treat.exp = "Placebo") \%>\%
  bind_rows(as_tibble(Pagliaro1992) \%>\% 
              select(id, treat.exp, bleed.exp, n.exp))

Pagliaro_new <- Pagliaro_new \%>\% 
  group_by(id) \%>\% 
  summarize(have_zero = (prod(bleed.exp) == 0) | (prod(n.exp - bleed.exp) ==0), 
            all_zero = (sum(bleed.exp) == 0) | (sum(n.exp - bleed.exp) ==0)) \%>\% 
  filter(!all_zero) \%>\%
  inner_join(Pagliaro_new) \%>\%
  mutate(y = log((bleed.exp + 0.5 * have_zero) / (n.exp - bleed.exp + 0.5 * have_zero)),
         v = 1/(bleed.exp + 0.5 * have_zero) + 1/(n.exp - bleed.exp + 0.5 * have_zero))

Pagliaro_res <- nma_incon(y = "y", v = "v", trt = "treat.exp", s = "id", 
                          comp = c("Sclerotherapy", "Beta-blocker"), data = Pagliaro_new)

}
\references{
Shih ...
Title.
\emph{Journal},
\bold{x}, p1--p2
}
\seealso{
\code{\link{nma}}, \code{\link{nma_side}}
}
\author{
Ming-Chieh Shih \email{mcshih@mx.nthu.edu.tw}
}
